engine/requirements.txt
flask
yfinance
pandas
psycopg2-binary
python-dotenv
pytz
gunicorn